GraphQL/transport_service/requirements.txt
fastapi
uvicorn[standard]
strawberry-graphql[fastapi]
